:hammer: fix: ProductDetail Page component 분리
